# docs/swagger.yaml
basePath: /api/v1
definitions:
  github_com_M1sterZag_Dont_Play_Separately_internal_core_transport_http_response.ErrorResponse:
    properties:
      error:
        type: string
      message:
        type: string
    type: object
  github_com_M1sterZag_Dont_Play_Separately_internal_core_transport_http_types.Nullable-string:
    properties:
      set:
        type: boolean
      value:
        type: string
    type: object
  internal_features_auth_transport_http.LoginRequest:
    properties:
      email:
        example: user@example.com
        type: string
      password:
        example: P@ssw0rd123!
        type: string
    required:
    - email
    - password
    type: object
...
  internal_features_auth_transport_http.TokenResponse:
    properties:
      access_token:
        type: string
      refresh_token:
        type: string
    type: object
...
  internal_features_games_transport_http.GetGameResponse:
    properties:
      icon_url:
        type: string
      id:
        type: integer
      slug:
        type: string
      title:
        type: string
    type: object
...
  internal_features_platforms_transport_http.PlatformDTOResponse:
    properties:
      abbreviation:
        type: string
      icon_url:
        type: string
      id:
        type: integer
      slug:
        type: string
      title:
        type: string
    type: object
...
  internal_features_teams_transport_http.GetTeamResponse:
    properties:
      contact_link:
        type: string
      created_at:
        type: string
      description:
        type: string
      desired_rating:
        type: string
      game_id:
        type: integer
      id:
        type: string
      is_active:
        type: boolean
      is_rating_required:
        type: boolean
      members:
        items:
          $ref: '#/definitions/internal_features_teams_transport_http.TeamMemberDTOResponse'
        type: array
      owner_id:
        type: string
      platform_id:
        type: integer
      slots_taken:
        type: integer
      slots_total:
        type: integer
      title:
        type: string
      version:
        type: integer
    type: object
  internal_features_teams_transport_http.PatchTeamRequest:
    properties:
      contact_link:
        $ref: '#/definitions/github_com_M1sterZag_Dont_Play_Separately_internal_core_transport_http_types.Nullable-string'
      description:
        $ref: '#/definitions/github_com_M1sterZag_Dont_Play_Separately_internal_core_transport_http_types.Nullable-string'
      desired_rating:
        $ref: '#/definitions/github_com_M1sterZag_Dont_Play_Separately_internal_core_transport_http_types.Nullable-string'
      is_rating_required:
        type: boolean
      slots_total:
        type: integer
      title:
        $ref: '#/definitions/github_com_M1sterZag_Dont_Play_Separately_internal_core_transport_http_types.Nullable-string'
    type: object
  internal_features_teams_transport_http.PatchTeamResponse:
    properties:
      contact_link:
        type: string
      created_at:
        type: string
      description:
        type: string
      desired_rating:
        type: string
      game_id:
        type: integer
      id:
        type: string
      is_active:
        type: boolean
      is_rating_required:
        type: boolean
      members:
        items:
          $ref: '#/definitions/internal_features_teams_transport_http.TeamMemberDTOResponse'
        type: array
      owner_id:
        type: string
      platform_id:
        type: integer
      slots_taken:
        type: integer
      slots_total:
        type: integer
      title:
        type: string
      version:
        type: integer
    type: object
...
  internal_features_teams_transport_http.TeamMemberDTOResponse:
    properties:
      nickname:
        type: string
      user_id:
        type: string
    type: object
...
host: 127.0.0.1:5050
info:
  contact: {}
  description: DPS Backend REST-API Schema
  title: DPS Backend API
  version: "1.0"
paths:
  /auth/login:
    post:
      consumes:
      - application/json
      description: Authenticates a user by email and password and returns an access
        and refresh token pair.
      parameters:
      - description: User credentials
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_features_auth_transport_http.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_features_auth_transport_http.TokenResponse'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/github_com_M1sterZag_Dont_Play_Separately_internal_core_transport_http_response.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_M1sterZag_Dont_Play_Separately_internal_core_transport_http_response.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/github_com_M1sterZag_Dont_Play_Separately_internal_core_transport_http_response.ErrorResponse'
      summary: Login
      tags:
      - auth
...
  /games/{id}:
    get:
      description: Returns a catalog game by its numeric ID.
      parameters:
      - description: Game ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_features_games_transport_http.GetGameResponse'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/github_com_M1sterZag_Dont_Play_Separately_internal_core_transport_http_response.ErrorResponse'
        "404":
          description: Not found
          schema:
            $ref: '#/definitions/github_com_M1sterZag_Dont_Play_Separately_internal_core_transport_http_response.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/github_com_M1sterZag_Dont_Play_Separately_internal_core_transport_http_response.ErrorResponse'
      summary: Get game by ID
      tags:
      - games
...
  /platforms/search:
    get:
      description: Searches the local catalog, falls back to IGDB if empty.
      parameters:
      - description: Search query
        in: query
        name: q
        required: true
        type: string
      - description: Page size
        in: query
        name: limit
        type: integer
      - description: Page offset
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/internal_features_platforms_transport_http.PlatformDTOResponse'
            type: array
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/github_com_M1sterZag_Dont_Play_Separately_internal_core_transport_http_response.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/github_com_M1sterZag_Dont_Play_Separately_internal_core_transport_http_response.ErrorResponse'
      summary: Search platforms
      tags:
      - platforms
...
  /teams/{team_id}:
    delete:
      description: Deletes a team. Only the owner can delete.
      parameters:
      - description: Team ID
        example: 330ca7c9-80bf-4808-a78b-d38cewer87b5
        format: uuid
        in: path
        name: team_id
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_M1sterZag_Dont_Play_Separately_internal_core_transport_http_response.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/github_com_M1sterZag_Dont_Play_Separately_internal_core_transport_http_response.ErrorResponse'
        "404":
          description: Not found
          schema:
            $ref: '#/definitions/github_com_M1sterZag_Dont_Play_Separately_internal_core_transport_http_response.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/github_com_M1sterZag_Dont_Play_Separately_internal_core_transport_http_response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete team
      tags:
      - teams
    get:
      description: Returns a team by ID together with its members.
      parameters:
      - description: Team ID
        example: 330ca7c9-80bf-4808-a78b-d38cewer87b5
        format: uuid
        in: path
        name: team_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_features_teams_transport_http.GetTeamResponse'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/github_com_M1sterZag_Dont_Play_Separately_internal_core_transport_http_response.ErrorResponse'
        "404":
          description: Not found
          schema:
            $ref: '#/definitions/github_com_M1sterZag_Dont_Play_Separately_internal_core_transport_http_response.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/github_com_M1sterZag_Dont_Play_Separately_internal_core_transport_http_response.ErrorResponse'
      summary: Get team
      tags:
      - teams
    put:
      consumes:
      - application/json
      description: 'Partially updates a team. Only the owner can update; pass "description":
        null to clear it.'
      parameters:
      - description: Team ID
        example: 330ca7c9-80bf-4808-a78b-d38cewer87b5
        format: uuid
        in: path
        name: team_id
        required: true
        type: string
      - description: Team patch payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_features_teams_transport_http.PatchTeamRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_features_teams_transport_http.PatchTeamResponse'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/github_com_M1sterZag_Dont_Play_Separately_internal_core_transport_http_response.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_M1sterZag_Dont_Play_Separately_internal_core_transport_http_response.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/github_com_M1sterZag_Dont_Play_Separately_internal_core_transport_http_response.ErrorResponse'
        "404":
          description: Not found
          schema:
            $ref: '#/definitions/github_com_M1sterZag_Dont_Play_Separately_internal_core_transport_http_response.ErrorResponse'
        "409":
          description: Conflict (optimistic lock)
          schema:
            $ref: '#/definitions/github_com_M1sterZag_Dont_Play_Separately_internal_core_transport_http_response.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/github_com_M1sterZag_Dont_Play_Separately_internal_core_transport_http_response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update team
      tags:
      - teams
  /teams/{team_id}/join:
    post:
      description: Adds the authenticated user to the team if it is active and has
        free slots.
      parameters:
      - description: Team ID
        example: 330ca7c9-80bf-4808-a78b-d38cewer87b5
        format: uuid
        in: path
        name: team_id
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/github_com_M1sterZag_Dont_Play_Separately_internal_core_transport_http_response.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_M1sterZag_Dont_Play_Separately_internal_core_transport_http_response.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/github_com_M1sterZag_Dont_Play_Separately_internal_core_transport_http_response.ErrorResponse'
        "404":
          description: Not found
          schema:
            $ref: '#/definitions/github_com_M1sterZag_Dont_Play_Separately_internal_core_transport_http_response.ErrorResponse'
        "409":
          description: Conflict (already a member / full / inactive)
          schema:
            $ref: '#/definitions/github_com_M1sterZag_Dont_Play_Separately_internal_core_transport_http_response.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/github_com_M1sterZag_Dont_Play_Separately_internal_core_transport_http_response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Join team
      tags:
      - teams
  /teams/{team_id}/leave:
    post:
      description: Removes the authenticated user from the team. The owner cannot
        leave the team.
      parameters:
      - description: Team ID
        example: 330ca7c9-80bf-4808-a78b-d38cewer87b5
        format: uuid
        in: path
        name: team_id
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/github_com_M1sterZag_Dont_Play_Separately_internal_core_transport_http_response.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_M1sterZag_Dont_Play_Separately_internal_core_transport_http_response.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/github_com_M1sterZag_Dont_Play_Separately_internal_core_transport_http_response.ErrorResponse'
        "404":
          description: Not found
          schema:
            $ref: '#/definitions/github_com_M1sterZag_Dont_Play_Separately_internal_core_transport_http_response.ErrorResponse'
        "409":
          description: Conflict (not a member / owner)
          schema:
            $ref: '#/definitions/github_com_M1sterZag_Dont_Play_Separately_internal_core_transport_http_response.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/github_com_M1sterZag_Dont_Play_Separately_internal_core_transport_http_response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Leave team
      tags:
      - teams
...
securityDefinitions:
  BearerAuth:
    description: Введите access-токен в формате "Bearer <token>"
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"